added graphics, optimization flags and fixed bug in mcts
Graphics with SDL2. Added -O2 as a flag in Makefile for compilation, to
make the search tree algorithm run faster. Fixed bug where if there was only
one move it would not be able to pick it second time it came to that node.
Fixed move notation so it is the opposite of the array indexed for y
cordinates (like it is on a proper board)